Made changes to chat app to enable docker deployment
