+++ MinuteMate/back/.env.example
@@ -1,0 +1,6 @@
+#OPENAI API KEY
+OPENAI_API_KEY = 
+
+# Weaviate cloud deployment
+WEAVIATE_URL = 
+WEAVIATE_API_KEY =
