+++ MinuteMate/back/Dockerfile
@@ -1,6 +1,7 @@
-FROM python:3.11-slim
-WORKDIR /App
-COPY . /App
-RUN python -m pip install -r requirements.txt
-EXPOSE 8000
-CMD [" ", "start","--port","8002","--host","0.0.0.0"]
+FROM python:3.11
+WORKDIR /app
+COPY . /app
+RUN pip install --upgrade pip
+RUN pip install -r requirements.txt
+EXPOSE 8001
+CMD ["uvicorn", "main:app","--host","0.0.0.0","--port","8001"]
