+++ MinuteMate/front/docker-compose.yml
@@ -1,0 +1,7 @@
+services:
+  chat_frontend:
+    build:
+      context: ./
+      dockerfile: Dockerfile
+    ports:
+      - 8501:8501
